requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
# openai==1.12.0
openai>=1.50.0
# httpx==0.24.1
httpx>=0.27.0
python-dotenv==1.0.0
redis
requests
anthropic>=0.25.0
python-multipart
boto3>=1.34.0